Centralize parquet object store helpers
