+++ volga_stream/src/runtime/functions/mod.rs
@@ -5,3 +5,4 @@ pub mod key_by;
 pub mod reduce;
 pub mod join;
 pub mod function_trait;
+pub mod parquet_utils;
