+++ volga_stream/src/runtime/functions/source/parquet/mod.rs
@@ -1,8 +1,7 @@
 use std::collections::HashMap;
-use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 
-use anyhow::{Result, anyhow};
+use anyhow::Result;
 use arrow::record_batch::RecordBatch;
 use arrow::datatypes::SchemaRef;
 use async_trait::async_trait;
@@ -15,6 +14,7 @@ use regex::Regex;
 
 use crate::common::message::Message;
 use crate::runtime::functions::function_trait::FunctionTrait;
+use crate::runtime::functions::parquet_utils::build_object_store;
 use crate::runtime::functions::source::source_function::SourceFunctionTrait;
 use crate::runtime::runtime_context::RuntimeContext;
 
@@ -75,7 +75,9 @@ impl SourceFunctionTrait for ParquetSourceFunction {
     async fn fetch(&mut self) -> Option<Message> {
         loop {
             if self.current_stream.lock().unwrap().is_none() {
-                let next_file = self.file_list.pop()?;
+                let Some(next_file) = self.file_list.pop() else {
+                    return None;
+                };
                 if let Ok(stream) = self.build_stream(&next_file).await {
                     *self.current_stream.lock().unwrap() = Some(stream);
                 } else {
@@ -106,7 +108,11 @@ impl SourceFunctionTrait for ParquetSourceFunction {
 #[async_trait]
 impl FunctionTrait for ParquetSourceFunction {
     async fn open(&mut self, ctx: &RuntimeContext) -> Result<()> {
-        let (store, prefix) = build_object_store(&self.config.spec.path, &self.config.spec.storage_options)?;
+        let (store, prefix) = build_object_store(
+            &self.config.spec.path,
+            &self.config.spec.storage_options,
+            true,
+        )?;
         let regex = self
             .config
             .spec
@@ -147,7 +153,11 @@ impl ParquetSourceFunction {
     }
 
     async fn build_stream(&self, meta: &ObjectMeta) -> Result<Box<dyn Stream<Item = Result<RecordBatch, parquet::errors::ParquetError>> + Unpin + Send>> {
-        let (store, _) = build_object_store(&self.config.spec.path, &self.config.spec.storage_options)?;
+        let (store, _) = build_object_store(
+            &self.config.spec.path,
+            &self.config.spec.storage_options,
+            true,
+        )?;
         let reader = ParquetObjectReader::new(store, meta.location.clone())
             .with_file_size(meta.size);
         let mut builder = ParquetRecordBatchStreamBuilder::new(reader).await?;
@@ -161,62 +171,6 @@ impl ParquetSourceFunction {
         }
         Ok(Box::new(builder.build()?))
     }
-}
-
-// Object store helpers
-fn build_object_store(
-    path: &str,
-    storage_options: &HashMap<String, String>,
-) -> Result<(Arc<dyn ObjectStore>, ObjectPath)> {
-    if path.starts_with("s3://") {
-        let (bucket, prefix) = split_s3_path(path)?;
-        let mut builder = object_store::aws::AmazonS3Builder::new()
-            .with_bucket_name(bucket);
-        if let Some(region) = storage_options.get("region") {
-            builder = builder.with_region(region);
-        }
-        if let Some(endpoint) = storage_options.get("endpoint_url") {
-            builder = builder.with_endpoint(endpoint);
-            if endpoint.starts_with("http://") {
-                builder = builder.with_allow_http(true);
-            }
-        }
-        if let (Some(key), Some(secret)) = (
-            storage_options.get("access_key_id"),
-            storage_options.get("secret_access_key"),
-        ) {
-            builder = builder.with_access_key_id(key).with_secret_access_key(secret);
-        }
-        let store = builder.build()?;
-        let store = Arc::new(store) as Arc<dyn ObjectStore>;
-        Ok((store, ObjectPath::from(prefix)))
-    } else {
-        let local_path = if let Some(stripped) = path.strip_prefix("file://") {
-            stripped
-        } else {
-            path
-        };
-        let local_path = PathBuf::from(local_path);
-        if local_path.is_file() {
-            let parent = local_path.parent().ok_or_else(|| anyhow!("invalid file path"))?;
-            let store = object_store::local::LocalFileSystem::new_with_prefix(parent)?;
-            let prefix = local_path.file_name().ok_or_else(|| anyhow!("invalid file path"))?.to_string_lossy().to_string();
-            let store = Arc::new(store) as Arc<dyn ObjectStore>;
-            Ok((store, ObjectPath::from(prefix)))
-        } else {
-            let store = object_store::local::LocalFileSystem::new_with_prefix(&local_path)?;
-            let store = Arc::new(store) as Arc<dyn ObjectStore>;
-            Ok((store, ObjectPath::from("")))
-        }
-    }
-}
-
-fn split_s3_path(path: &str) -> Result<(String, String)> {
-    let stripped = path.trim_start_matches("s3://");
-    let mut parts = stripped.splitn(2, '/');
-    let bucket = parts.next().ok_or_else(|| anyhow!("missing s3 bucket"))?.to_string();
-    let prefix = parts.next().unwrap_or("").to_string();
-    Ok((bucket, prefix))
 }
 
 async fn list_objects(
